Add base language server configuration
This commit adds the base language server configuration for
NeoVim. We are using the ready made `lsp-zeror` for now and
as we go we can adopt it when necessary.

This commit also couple of the core plugins I have been using
on vim, and it also configures those to keep the same binding
and keymaps.